+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Adds some small styling adjustments to the hoursCollapsible - kieran
 - Fix overfetching data on Fair Detail - luc
 - Update Show detail view spacing & copy - luc
+- Adds missing fair booth separator - kieran
 
 ### 1.8.1